+++ swift_browser_ui/upload/common.py
@@ -2,10 +2,17 @@
 
 
 import typing
+import logging
+import os
 
 import aiohttp.web
 
+import swift_browser_ui.upload.cryptupload as cryptupload
 from swift_browser_ui.upload import upload
+
+
+LOGGER = logging.getLogger(__name__)
+LOGGER.setLevel(os.environ.get("LOG_LEVEL", "INFO"))
 
 
 DATA_PREFIX = "data/"
@@ -104,6 +111,40 @@ async def get_upload_instance(
         await upload_session.a_check_container()
         request.app[session]["uploads"][pro][cont][ident] = upload_session
 
+    return upload_session
+
+
+async def get_encrypted_upload_instance(
+    request: aiohttp.web.Request,
+) -> cryptupload.EncryptedUploadProxy:
+    """Return the specific encrypted upload proxy for the object."""
+    session = get_session_id(request)
+
+    project = request.match_info["project"]
+    container = request.match_info["container"]
+    object_name = request.match_info["object_name"]
+
+    try:
+        upload_session = request.app[session]["enuploads"][project][container][
+            object_name
+        ]
+        LOGGER.info("Returning an existing upload session.")
+    except KeyError:
+        LOGGER.info("Creating a new upload session.")
+        # Check the existence of the dictionary structure
+        if project not in request.app[session]["enuploads"]:
+            request.app[session]["enuploads"][project] = {}
+        if container not in request.app[session]["enuploads"][project]:
+            request.app[session]["enuploads"][project][container] = {}
+        if object_name not in request.app[session]["enuploads"][project][container]:
+            upload_session = cryptupload.EncryptedUploadProxy(
+                request.app[session],
+                request.app["client"],
+            )
+            request.app[session]["enuploads"][project][container][
+                object_name
+            ] = upload_session
+    LOGGER.info(f"Session object id: {id(upload_session)}")
     return upload_session